2 forgotten tests move to integration
